AegisRAG Intelligence Agent/src/common.py
import os
from dotenv import load_dotenv

# Load environment variables from .env
load_dotenv()

# OpenAI
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

# Chroma configuration
CHROMA_DIR = os.getenv("CHROMA_DIR", "./chroma_db")
CHROMA_COLLECTION = os.getenv("CHROMA_COLLECTION", "research_corpus")
